Ribbon Font group 2-row redesign, selection-offset fix, finer scrolling, Mermaid rendering
- Ribbon Font group rebuilt LibreOffice-style: new AP_RIBBON_ITEM_ROWEND
  packs groups row-major (per-row GtkBox) so narrow glyph buttons no
  longer share grid columns with the wide font combo. Row 1: font +
  size combos, Grow/Shrink Font, Clear Formatting; row 2: B/I/U/S/
  x2/x2 glyph strip, highlight + font-colour glyph buttons, Font dialog.
- AP_RIBBON_FLAG_GLYPH draws Pango glyphs (bold B, italic I, underlined
  U, struck S, x2/x2, A+/A-); colour buttons get descriptive glyphs;
  icon-only buttons fall back to text when the theme lacks the icon.
- New menu items: Format > Text > Grow/Shrink Font + Clear Formatting
  (new clearFormatting edit method -> FV_View::resetCharFormat), Insert
  > Edit Equation, Help > Credits - every ribbon button now resolves to
  a real GAction; dead Split Table entry removed from the ribbon.
- Fixed click/drag selection offset: gdk_event_get_position() returns
  surface-relative coords under GTK4 (~157px off); EV_UnixMouse now
  uses the gesture callbacks' widget-relative x/y, scroll position is
  translated via gtk_widget_compute_point.
- Finer vertical scrolling: wheel step 60px -> 36px, smooth touchpad
  deltas scroll proportionally with fractional accumulation, and
  vScrollChanged coalesces pending scroll targets instead of dropping
  them.
- Mermaid fenced blocks in Markdown render to embedded PNGs via a new
  built-in Cairo renderer (flowchart, sequence, Gantt, class, pie).

+++ src/text/fmt/xp/fp_PageSize.cpp
@@ -7,6 +7,7 @@
 
 #include "string.h"
 #include <math.h>
+#include <stdint.h>
 
 #if defined(__GLIBC__)
 #include <langinfo.h>
@@ -161,9 +162,10 @@ const char * fp_PageSize::getDefaultPageName(void)
 {
 #ifdef ABI_HAVE_NL_PAPER_SIZE
 	// _NL_PAPER_WIDTH/_NL_PAPER_HEIGHT honor LC_PAPER; nl_langinfo
-	// returns these integer items cast to a pointer (LP#234756)
-	const long w = (long) nl_langinfo(_NL_PAPER_WIDTH);
-	const long h = (long) nl_langinfo(_NL_PAPER_HEIGHT);
+	// returns these integer items cast to a pointer (LP#234756).
+	// Read them as int: the upper pointer bits are undefined.
+	const int w = (int)(intptr_t) nl_langinfo(_NL_PAPER_WIDTH);
+	const int h = (int)(intptr_t) nl_langinfo(_NL_PAPER_HEIGHT);
 	if (w > 0 && h > 0)
 	{
 		for (int i = 0; i < _last_predefined_pagesize_dont_use_; i++)
